re #1010 casting to int so numerical value is written to stream

+++ PlusLib/src/PlusDataCollection/Philips3D/vtkPlusPhilips3DProbeVideoSource.cxx
@@ -377,7 +377,7 @@ void vtkPlusPhilips3DProbeVideoSource::CallbackAddFrame(vtkImageData* imageData)
   });
 
   addTask.wait();
-  uint8_t maxPixelValue = maxPixelTask.get();
+  unsigned int maxPixelValue = static_cast<unsigned int>(maxPixelTask.get());
 
   std::stringstream ss;
   ss << maxPixelValue;
